+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#141] configuration keys used for cas password-policies are also applied to redmine
 
 ## [v5.1.4-1] - 2024-12-16
 - [#138] Update Redmine to version 5.1.4