+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v5.1.4-2] - 2025-01-09
+### Added
+- [#141] configuration keys used for cas password-policies are also applied to redmine
+
 ## [v5.1.4-1] - 2024-12-16
 - [#138] Update Redmine to version 5.1.4